refactor: de-part game_map_area state logic clusters for wave 9 slice E (#4117)
Merge four part-based map_state libraries into explicit single-file modules
(state logic, province actions, civilian/fleet draft projection) and shrink
the no-part grandfather allowlist by 15 entries while staying within the
main-line budget.

Refs #4117
