+++ app/lib/features/game/flame/map_state/game_map_area_province_action_states.dart
@@ -4,9 +4,9 @@ import 'package:colonizethis_logic/colonizethis_logic.dart';
 import 'package:colonizethis_models/colonizethis_models.dart' as ct_models;
 import 'package:colonizethis_map/colonizethis_map.dart';
 
-part 'game_map_area_province_action_states_prospect.dart';
-part 'game_map_area_province_action_states_explore.dart';
-part 'game_map_area_province_action_states_build_improvement.dart';
+
+
+
 
 /// Province-overlay action visibility/enablement computations for prospect,
 /// explore, and build-improvement shortcuts.
@@ -108,4 +108,302 @@ class GameMapAreaProvinceActionStates {
         currentOrders: currentOrders,
         tileMapByRegion: tileMapByRegion,
       );
+}
+
+/// Prospect inline-action visibility/enablement for province overlay.
+abstract final class GameMapAreaProvinceActionStatesProspect {
+  /// Returns province-overlay prospect action visibility + enablement.
+  ///
+  /// The panel must read stable world/player tile state only so map scrolling
+  /// and rebuild churn do not trigger expensive order-engine validation.
+  static ({bool showIcon, bool enabled, bool hasExplorerUnits}) compute({
+    required ct_models.Game game,
+    required String humanPlayerId,
+    required String selectedTileKey,
+    required PlayerView playerView,
+    required MapTopology? topology,
+    required ct_models.Orders currentOrders,
+    required Map<String, TileMapResult>? tileMapByRegion,
+  }) {
+    final parsed = tryParseTileKey(selectedTileKey);
+    if (parsed == null) {
+      return (showIcon: false, enabled: false, hasExplorerUnits: false);
+    }
+    final tileVisibility = playerView.visibilityForTile(selectedTileKey);
+    if (tileVisibility == VisibilityLevel.unknown) {
+      return (showIcon: false, enabled: false, hasExplorerUnits: false);
+    }
+    final prefixedProvinceId = parsed.prefixedProvinceId;
+    final isProvinceTile =
+        game.worldState.tryGetProvince(prefixedProvinceId) != null;
+    if (!isProvinceTile) {
+      return (showIcon: false, enabled: false, hasExplorerUnits: false);
+    }
+
+    final isMineralEligible = isMineralEligibleTile(
+      game,
+      tileMapByRegion,
+      selectedTileKey,
+    );
+    if (!isMineralEligible) {
+      return (showIcon: false, enabled: false, hasExplorerUnits: false);
+    }
+
+    final playerProspectedTiles =
+        game.worldState.playerProspectedTiles[humanPlayerId] ??
+        const <String>{};
+    if (playerProspectedTiles.contains(selectedTileKey)) {
+      return (showIcon: false, enabled: false, hasExplorerUnits: false);
+    }
+
+    final allUnits = <ct_models.Unit>[
+      ...game.worldState.oldWorld.units,
+      ...game.worldState.newWorld.units,
+    ];
+    final explorerUnits = allUnits
+        .where((unit) => unit.ownerId == humanPlayerId)
+        .where(
+          (unit) =>
+              workOrderTargetsByUnitType[unit.type]?.contains(
+                kWorkTargetProspect,
+              ) ??
+              false,
+        )
+        .toList();
+    if (explorerUnits.isEmpty) {
+      return (showIcon: true, enabled: false, hasExplorerUnits: false);
+    }
+    // Refs #3753 R4/R4b: prospect inside a Minor/Tribe province requires a
+    // Consulate (or higher). The prospect enablement does not flow through the
+    // work-order validator, so apply the shared Consulate gate here to keep the
+    // inline action disabled (rather than enabled-then-rejected) — the overlay
+    // surfaces the rejection tooltip.
+    final prospectOwnerId = game.worldState
+        .tryGetProvince(prefixedProvinceId)
+        ?.ownerId;
+    if (explorerConsulateGateBlocksMinorTribeProvince(
+      game: game,
+      playerId: humanPlayerId,
+      provinceOwnerId: prospectOwnerId,
+    )) {
+      return (showIcon: true, enabled: false, hasExplorerUnits: true);
+    }
+    return (showIcon: true, enabled: true, hasExplorerUnits: true);
+  }
+}
+
+/// Explore inline-action visibility/enablement for province overlay.
+abstract final class GameMapAreaProvinceActionStatesExplore {
+  static Set<String> buildEligibleTileKeyCache({
+    required ct_models.Game game,
+    required String humanPlayerId,
+    required PlayerView playerView,
+    required MapTopology topology,
+    required Map<String, TileMapResult>? tileMapByRegion,
+    required ct_models.Orders currentOrders,
+  }) {
+    final cache = PerPlayerWorkTargetSelectionCache();
+    cache.refresh(
+      WorkTargetSelectionSnapshot(
+        game: game,
+        playerId: humanPlayerId,
+        playerView: playerView,
+        topology: topology,
+        currentOrders: currentOrders,
+        tileMapByRegion: tileMapByRegion,
+      ),
+    );
+    return cache.get(humanPlayerId, kWorkTargetExplore);
+  }
+
+  static ({bool showIcon, bool enabled, bool hasExplorerUnits}) compute({
+    required ct_models.Game game,
+    required String humanPlayerId,
+    required String selectedTileKey,
+    required RegionMapViewData selectedRegion,
+    PerPlayerWorkTargetSelectionCache? workTargetSelectionCache,
+    Set<String>? cachedExploreEligibleTileKeys,
+  }) {
+    final parsed = tryParseTileKey(selectedTileKey);
+    if (parsed == null || parsed.regionId != selectedRegion.regionId) {
+      return GameMapAreaProvinceActionStates.kHiddenExplorerInlineActionState;
+    }
+    final tileProvinceId = parsed.provinceLocalId;
+    final prefixedProvinceId = parsed.prefixedProvinceId;
+    final province = game.worldState.tryGetProvince(prefixedProvinceId);
+    if (province == null) {
+      return GameMapAreaProvinceActionStates.kHiddenExplorerInlineActionState;
+    }
+
+    final x = parsed.x;
+    final y = parsed.y;
+    if (x < 0 ||
+        y < 0 ||
+        x >= selectedRegion.width ||
+        y >= selectedRegion.height) {
+      return GameMapAreaProvinceActionStates.kHiddenExplorerInlineActionState;
+    }
+    final selectedCell = selectedRegion.cellAt(x, y);
+    if (selectedCell.visibility == TileVisibility.unrevealed) {
+      return GameMapAreaProvinceActionStates.kHiddenExplorerInlineActionState;
+    }
+
+    final provinceCells = selectedRegion.cells
+        .where((cell) => !cell.isSea && cell.regionCellId == tileProvinceId)
+        .toList();
+    if (provinceCells.isEmpty) {
+      return GameMapAreaProvinceActionStates.kHiddenExplorerInlineActionState;
+    }
+    final hasUnrevealed = provinceCells.any(
+      (cell) => cell.visibility == TileVisibility.unrevealed,
+    );
+    final hasRevealed = provinceCells.any(
+      (cell) => cell.visibility != TileVisibility.unrevealed,
+    );
+    if (!hasUnrevealed || !hasRevealed) {
+      return GameMapAreaProvinceActionStates.kHiddenExplorerInlineActionState;
+    }
+
+    final allUnits = <ct_models.Unit>[
+      ...game.worldState.oldWorld.units,
+      ...game.worldState.newWorld.units,
+    ];
+    final hasExplorerUnits = allUnits
+        .where((unit) => unit.ownerId == humanPlayerId)
+        .any(
+          (unit) =>
+              workOrderTargetsByUnitType[unit.type]?.contains(
+                kWorkTargetExplore,
+              ) ??
+              false,
+        );
+
+    final eligibleTileKeys =
+        cachedExploreEligibleTileKeys ??
+        workTargetSelectionCache?.get(humanPlayerId, kWorkTargetExplore) ??
+        const <String>{};
+    final hasEligibleExploreTarget = eligibleTileKeys.any((tileKey) {
+      final p = tryParseTileKey(tileKey);
+      return p != null &&
+          p.regionId == selectedRegion.regionId &&
+          p.provinceLocalId == tileProvinceId;
+    });
+    if (!hasEligibleExploreTarget) {
+      // Refs #3753 R4/R4b: when the only blocker is the Consulate gate (a known
+      // Minor/Tribe province with no Consulate filters out every eligible
+      // explore tile), show the inline action disabled (not hidden) so the
+      // overlay can surface the rejection tooltip. Other ineligibility reasons
+      // keep the icon hidden as before.
+      final ownerId = game.worldState.tryGetProvince(prefixedProvinceId)?.ownerId;
+      if (explorerConsulateGateBlocksMinorTribeProvince(
+        game: game,
+        playerId: humanPlayerId,
+        provinceOwnerId: ownerId,
+      )) {
+        return (
+          showIcon: true,
+          enabled: false,
+          hasExplorerUnits: hasExplorerUnits,
+        );
+      }
+      return GameMapAreaProvinceActionStates.kHiddenExplorerInlineActionState;
+    }
+
+    return (
+      showIcon: true,
+      enabled: hasExplorerUnits,
+      hasExplorerUnits: hasExplorerUnits,
+    );
+  }
+}
+
+/// Build-improvement inline-action visibility/enablement for province overlay.
+abstract final class GameMapAreaProvinceActionStatesBuildImprovement {
+  static ({bool showIcon, bool enabled, bool hasBuilderUnits}) compute({
+    required ct_models.Game game,
+    required String humanPlayerId,
+    required String selectedTileKey,
+    required PlayerView playerView,
+    PerPlayerWorkTargetSelectionCache? workTargetSelectionCache,
+    MapTopology? topology,
+    ct_models.Orders currentOrders = const ct_models.Orders(),
+    Map<String, TileMapResult>? tileMapByRegion,
+  }) {
+    final parsed = tryParseTileKey(selectedTileKey);
+    if (parsed == null) {
+      return GameMapAreaProvinceActionStates.kHiddenBuilderInlineActionState;
+    }
+    final tileVisibility = playerView.visibilityForTile(selectedTileKey);
+    if (tileVisibility == VisibilityLevel.unknown) {
+      return GameMapAreaProvinceActionStates.kHiddenBuilderInlineActionState;
+    }
+    final prefixedProvinceId = parsed.prefixedProvinceId;
+    final isProvinceTile =
+        game.worldState.tryGetProvince(prefixedProvinceId) != null;
+    if (!isProvinceTile) {
+      return GameMapAreaProvinceActionStates.kHiddenBuilderInlineActionState;
+    }
+    final player = game.playerById(humanPlayerId);
+    if (player == null) {
+      return GameMapAreaProvinceActionStates.kHiddenBuilderInlineActionState;
+    }
+
+    final resourceId = game.worldState.resourceByTileKey[selectedTileKey];
+    if (resourceId == null || resourceId.isEmpty) {
+      return GameMapAreaProvinceActionStates.kHiddenBuilderInlineActionState;
+    }
+    final currentLevel = game.worldState.tileState.improvementLevel(
+      selectedTileKey,
+    );
+    final techCap = extractionCapForResourceForUnlocked(
+      player.techUnlocked,
+      resourceId,
+    );
+    final terrain = terrainTypeForTileKey(tileMapByRegion, selectedTileKey);
+    final effectiveCap = terrain == null
+        ? techCap
+        : clampExtractionCapForTerrain(techCap, resourceId, terrain);
+    if (currentLevel >= effectiveCap) {
+      return GameMapAreaProvinceActionStates.kHiddenBuilderInlineActionState;
+    }
+
+    final allUnits = <ct_models.Unit>[
+      ...game.worldState.oldWorld.units,
+      ...game.worldState.newWorld.units,
+    ];
+    final builderUnits = allUnits
+        .where((unit) => unit.ownerId == humanPlayerId)
+        .where(
+          (unit) =>
+              workOrderTargetsByUnitType[unit.type]?.contains(
+                kWorkTargetBuildImprovement,
+              ) ??
+              false,
+        )
+        .toList();
+    if (builderUnits.isEmpty) {
+      return (showIcon: true, enabled: false, hasBuilderUnits: false);
+    }
+    final anyAssignable =
+        workTargetSelectionCache?.contains(
+          humanPlayerId,
+          kWorkTargetBuildImprovement,
+          selectedTileKey,
+        ) ??
+        (topology == null
+            ? false
+            : builderUnits.any((builder) {
+                final valid = getValidWorkOrderTileKeysWithVisibility(
+                  game: game,
+                  topology: topology,
+                  view: playerView,
+                  unitId: builder.id,
+                  workTarget: kWorkTargetBuildImprovement,
+                  currentOrders: currentOrders,
+                  tileMapByRegion: tileMapByRegion,
+                );
+                return valid.contains(selectedTileKey);
+              }));
+    return (showIcon: true, enabled: anyAssignable, hasBuilderUnits: true);
+  }
 }
